rename transformer -> transformer_decoder
